refactor: Move reusable types into types/ dir
