+++ react-frontend/src/components/Logout.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { Navigate } from "react-router-dom";
-import type { LoginStatusProps } from "./LoginStatusProps";
+import type { LoginStatusProps } from "../types/LoginStatusProps";
 
 
 function Logout({setLoginStatus}: LoginStatusProps) {
