+++ react-frontend/src/types/Recipe.tsx
@@ -1,0 +1,10 @@
+interface Recipe {
+    id: string,
+    title: string,
+    description: string,
+    source: string,
+    instructions: string[],
+    ingredientInfos: string[]
+}
+
+export type { Recipe };
